ref="unique_id"
it's a attribute in html but only known to vue, it somewhat works like id and we can extraction the whole element from the script like this.$refs.unique_id.

It helps us to use value when we need rather than every keystroke and handling the whole element is easy.